# Conflicts:
#	KNUCinema/src/main/java/com/example/KNUCinema/KnuCinemaController.java

+++ KNUCinema/src/main/java/com/example/KNUCinema/ReservationDTO.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class ReservationDTO {
     private long id;
     private CinemaDTO cinema;
-    private int UserId;
+    private long UserId;
 }
 
 //TODO : id는 자동으로 set금지 하는 어노테이션
